Fixed bug that could cause track length to be written incorrectly
The pointer midTrkLen to malloc:ed data was reused after the data had
beed realloc:ed. This sometimes results in the wrong memory location being
written and the track length header field to be incorrect.

The track length in a MIDI file is redundant since each track is in any
case concluded with an End-of-track event. Software can therefore ignore
the track length header field. But Windows Media Player will refuse to
play MIDI files where the track length header does not match the actual
length.